fix: move styles folder to new location
