+++ components/Pokemon/Styles.ts
